fix: remaining safe exports
