+++ packages/core/src/index.ts
@@ -18,6 +18,7 @@ export * from "./util/date";
 export * from "./util/firmware";
 export * from "./util/graph";
 export * from "./util/misc";
+export * from "./util/_Types";
 export * from "./values/Cache";
 export * from "./values/CacheBackedMap";
 export * from "./values/Duration";
@@ -25,3 +26,4 @@ export * from "./values/Metadata";
 export * from "./values/Primitive";
 export * from "./values/Timeout";
 export * from "./values/ValueDB";
+export * from "./values/_Types";
