+++ packages/core/src/index_safe.ts
@@ -12,13 +12,13 @@ export * from "./security/DSK";
 export * from "./security/SecurityClass";
 export * from "./test/assertZWaveError";
 export * from "./util/crc";
-export * from "./util/firmware";
 export * from "./util/graph";
 export * from "./util/misc";
+export * from "./util/_Types";
 export * from "./values/Cache";
 export * from "./values/CacheBackedMap";
 export * from "./values/Duration";
 export * from "./values/Metadata";
 export * from "./values/Primitive";
 export * from "./values/Timeout";
-export * from "./values/ValueDB";
+export * from "./values/_Types";
