+++ packages/core/src/util/_Types.ts
@@ -1,0 +1,13 @@
+export type FirmwareFileFormat =
+	| "aeotec"
+	| "otz"
+	| "ota"
+	| "hex"
+	| "hec"
+	| "gecko"
+	| "bin";
+
+export interface Firmware {
+	data: Buffer;
+	firmwareTarget?: number;
+}
